backend/requirements.txt
aiosmtplib==2.0.1
alembic==1.10.3
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
fastapi-mail==1.2.6
flake8==6.0.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.0
sqlparse==0.4.3
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.14.0
